modificaciones

+++ src/gyms/gyms.service.ts
@@ -2,21 +2,49 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateGymDto } from './dto/create-gym.dto';
 import { UpdateGymDto } from './dto/update-gym.dto';
 import { GymsRepository } from './gyms.repository';
+import axios from 'axios';
+import { JwtService } from '@nestjs/jwt';
+import { Roles } from 'src/Roles/roles.enum';
+import { JWT_SECRET } from 'src/config/env.config';
 
 @Injectable()
 export class GymsService {
   constructor(
     private readonly gymsRepository: GymsRepository,
-    private readonly gymsService: GymsService
+    private readonly jwtService: JwtService
   ) {}
 
-  async create(createGymDto: CreateGymDto) {
+  async create(token: string, createGymDto: CreateGymDto) {
     try{
+      const decoded = this.jwtService.decode(token);
+      const owner = decoded.id
       const slug = await this.gymsRepository.findBySlug(createGymDto.slug);
+
       if (slug) {
         throw new BadRequestException('Slug already exists');
       }
-      return this.gymsRepository.create(createGymDto);
+
+      await this.gymsRepository.create(owner, createGymDto);
+
+      const patchRole = this.jwtService.sign({
+        role: Roles.Admin
+      }, {secret: JWT_SECRET})
+      const response = await axios.patch('http://localhost:3001/auth/role', 
+        {
+          email: decoded.email,
+          role: Roles.Admin
+        },
+        {
+          headers: { 'Authorization': `Bearer ${patchRole}` }
+        }
+      );
+
+      const newToken = response.data
+      const newGym = await this.gymsRepository.findBySlug(createGymDto.slug);
+      console.log(newGym);
+      
+      return { gym: newGym, token: newToken.newToken };
+
     } catch (error) {
       console.log(error);
       
@@ -49,7 +77,7 @@ export class GymsService {
 
   async update(id: string, updateGymDto: UpdateGymDto) {
     try{
-      const gym = await this.gymsService.findById(id);
+      const gym = await this.gymsRepository.findById(id);
       if (!gym) {
         throw new BadRequestException('Gym not found');
       }
@@ -59,9 +87,23 @@ export class GymsService {
     }
   }
 
-  async remove(id: string) {
+  async remove(id: string, token: string) {
     try{
+      const decoded = this.jwtService.decode(token);
+      const gym = await this.gymsRepository.findById(id);
+      if (decoded.id !== gym.owner) {
+        throw new BadRequestException('You are not the owner of this gym');
+      }
       await this.gymsRepository.remove(id);
+      await axios.patch('http://localhost:3001/auth/role', 
+        {
+          email: decoded.email,
+          role: Roles.User
+        },
+        {
+          headers: { 'Authorization': `Bearer ${token}` }
+        }
+      );
     } catch (error) {
       console.log(error);
     }
